pyproject.toml
[tool.poetry]
name = "exclamationlifebot"
version = "0.1.0"
description = ""
authors = ["SirCypkowskyy <dcyprian.a.gburek@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.21"
python-dotenv = "^0.20.0"
beautifulsoup4 = "^4.11.1"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
